requirements.txt
streamlit==1.39.0
pandas==2.2.3
requests==2.32.3
supabase==2.6.0
python-dateutil==2.9.0.post0
matplotlib==3.9.2
streamlit==1.39.0
pandas==2.2.3
requests==2.32.3
supabase==2.6.0
python-dateutil==2.9.0.post0
matplotlib==3.9.2
pyarrow==16.1.0
